Task 3. API tests refactoring
